Fix SA1402 violation in SlideData.cs
